pop up rendering and .js file

+++ FlightReservation/src/main/java/com/pnwairlines/flightreservation/controllers/SeatController.java
@@ -46,7 +46,8 @@ public class SeatController {
 				Model model
 			) {	
 				Seat thisSeat = seatServ.getOne(id);
-				Double total = Math.ceil(thisSeat.getPrice() * 1.10);
+				
+				int total = (int) Math.ceil(thisSeat.getPrice() * 1.10);
 				session.setAttribute("seat_id", id);
 				session.setAttribute("total", total);
 				model.addAttribute("seat", seatServ.getOne(id));
